requirements.txt
streamlit>=1.0.0
pandas>=1.0.0
scikit-learn>=0.24.0
liac-arff>=2.5.0